--- src/message.ts
@@ -1,4 +1,0 @@
-export interface Message {
-    title: string;
-    content: string;
-}
