--- src/messageBuilder/iMessageBuilder.ts
@@ -1,5 +1,0 @@
-import { Message } from "../message";
-
-export interface MessageBuilder {
-    buildMessage(): Promise<Message> 
-}
